// libraries/common-ui/src/types.d.ts
export type ChartValue = string | number | Date;
// ...
export type Aggregation =
  | "count"
  | "sum"
  | "min"
  | "max"
  | "avg"
  | "median"
  | "deviation"
  | "variance"
  | "quantileThree"
  | "quantileOne";
// ...
export interface IChartValue {
  idx: number;
  keyData: ChartValue;
  valueData: number;
  valueField: string;
  outlier?: boolean;
  highlight?: boolean;
}

export interface IChartAggregation {
  valueData: number;
  valueField: string;
}